PS Subtyped

PS Subtyped

Usage

ps_subtyped_stan(
  num_data,
  num_knots,
  num_path,
  knots,
  spline_degree,
  Y,
  P1,
  P2,
  X,
  week_effect,
  DOW,
  cov_structure = c(0, 1, 2),
  noise_structure = c(0, 1),
  ...
)

Arguments

num_data

number of data points

num_knots

number of knots

num_path

number of pathogens

knots

the sequence of knots

spline_degree

the degree of spline (is equal to order - 1)

Y

outcome variable, eg daily number of cases

P1

daily number of lab tests positive for influenza A (must be 1st column) and all other pathogens

P2

daily number of influenza A subtypes eg H3N2, H1N1

X

time data

week_effect

Number of distinct days in day of week effect. 1 = single effect shared by all days (essentially no DOW), 7 = each day a unique effect

DOW

integer of day of the week

cov_structure

0 is tau of 1, 1 is tau of num_path, 2 is Sigma with two dimensions: num_path, num_path.

noise_structure

0 only includes observation noise (same between pathogens), 1 includes noise in individual pathogens as well

...

Arguments passed to `rstan::sampling` (e.g. iter, chains).

Value

An object of class `stanfit` returned by `rstan::sampling`