Single pathogen structure

Single pathogen structure

Usage

single(data, case_timeseries, time = NULL)

Arguments

data

dataframe containing columns with all relevant data, or a time series object (ts, xts, zoo, tsibble, etc.)

case_timeseries

Column name containing case counts. Must be numeric or a units object from the units package.

time

name of column with time data. Required for non-time-series input data. Flexible format - can be date, index, or others, accepted as index identifiers in the tsibble time format. Optional when data is a time series class object (ts, mts, xts, zoo, zooreg, tsibble) as the time index will be automatically detected.

Value

formatted list with pathogen structure and data of class EpiStrainDynamics.pathogen_structure.

See also

Examples

# Using a data frame
single(
  data = sarscov2,
  case_timeseries = 'cases',
  time = 'date'
)
#> $pathogen_structure
#> [1] "single"
#> 
#> $pathogen_names
#> [1] "cases"
#> 
#> $validated_tsbl
#> # A tsibble: 830 x 2 [1D]
#>    date       case_timeseries
#>    <date>               <dbl>
#>  1 2020-09-23            7096
#>  2 2020-09-24            7682
#>  3 2020-09-25            7318
#>  4 2020-09-26            6810
#>  5 2020-09-27            7023
#>  6 2020-09-28            9983
#>  7 2020-09-29           10366
#>  8 2020-09-30           12717
#>  9 2020-10-01           13357
#> 10 2020-10-02           13854
#> # ℹ 820 more rows
#> 
#> $data
#> $data$case_timeseries
#>   [1]   7096   7682   7318   6810   7023   9983  10366  12717  13357  13854
#>  [11]  11667  11953  16753  17250  18573  18425  15893  12648  12202  19671
#>  [21]  19116  19920  18606  17915  14882  14311  25872  25613  25714  23475
#>  [31]  21694  16434  15895  26807  24351  23901  23614  22890  16673  15970
#>  [41]  31925  25943  24027  24063  23659  18766  20244  31510  27789  27757
#>  [51]  24826  24388  18223  16233  26952  23488  21076  18041  17100  12533
#>  [61]  11895  19024  16679  16814  14761  14554  11086  10852  18198  16684
#>  [71]  16676  15860  15834  12893  13731  21584  21065  21909  22916  24390
#>  [81]  19588  22266  35197  34378  35427  34437  36419  25589  33066  48192
#>  [91]  46475  42936  32392  14419  41110  48520  46408  83090  72161  53947
#> [101]  32482  61654  56335  77972  66638  58784  53187  48105  40035  37174
#> [111]  58592  50647  46509  43472  41452  31250  29592  45957  40525  35986
#> [121]  32502  30330  22344  17554  30737  27730  26303  24773  22434  17167
#> [131]  15783  22972  20265  19914  18493  16307  12320  11664  16194  14225
#> [141]  13899  13139  12573   9035   8995  14735  12939  11942  11793  10728
#> [151]   8313   8098  11657   9763   8847   8024   6782   4949   4873   7192
#> [161]   6918   6172   5937   5467   4628   4410   6890   6488   6250   5933
#> [171]   5717   4381   4425   6703   5863   5978   5632   5180   4247   5390
#> [181]   6459   5484   6224   5616   4567   3731   5030   4196   3779   3966
#> [191]   3674   2747   2508   2412   2744   3313   3253   3147   2953   2380
#> [201]   2376   2881   2495   2512   2345   2347   1722   2584   2738   2541
#> [211]   2708   2409   2020   1564   2119   2626   2178   2499   2230   1825
#> [221]   1433   1521   2084   2553   2569   2326   1982   1609   2307   2809
#> [231]   2329   2335   2182   1938   1628   1923   2858   2812   2813   2517
#> [241]   2367   2000   2326   3369   3427   3922   3747   3589   3000   3164
#> [251]   3897   5431   6168   6258   5861   4731   5534   7946   7896   8210
#> [261]   8043   7619   6284   7135  10362  10800  11052  11106  10320   8447
#> [271]  10087  16308  17402  18351  18714  18169  16301  18502  28418  28199
#> [281]  28923  28794  26339  24010  25170  34919  36060  38597  34020  32804
#> [291]  28379  30503  44511  47707  54910  62000  55264  38476  36182  47924
#> [301]  39876  35548  31278  27817  22852  22187  29981  29582  30237  28432
#> [311]  24749  20570  20985  29840  30473  32338  31660  28035  22540  22887
#> [321]  32162  31518  33350  31726  28974  25057  25327  36907  37920  38739
#> [331]  36501  32231  28536  28743  39715  37682  38777  37613  33745  27601
#> [341]  26136  34774  45212  43532  42328  38633  31810  31803  44929  38619
#> [351]  35849  33974  29187  23533  24060  34658  31435  30414  29710  29479
#> [361]  26645  31084  40768  38848  36756  35851  32108  28567  32083  41589
#> [371]  36548  34837  32604  29058  26316  32059  43733  39279  39264  38465
#> [381]  35612  32306  36418  50304  48321  47775  46419  43560  37878  43752
#> [391]  57743  52529  51661  45194  40664  35103  35764  47466  42124  43045
#> [401]  41391  40251  33569  34814  45109  38019  35986  33662  30318  27436
#> [411]  30586  44126  42246  42942  40999  37763  32885  35532  50948  46430
#> [421]  45895  44062  39606  35335  37250  50733  48197  46772  43425  39330
#> [431]  33303  39706  56289  55695  54820  50319  45667  38784  41051  59694
#> [441]  56258  58827  56995  52492  48594  53158  87169 102788 112703 106055
#> [451]  95897  88883  89133 133966 150293 161560 160633 137268  70363 116287
#> [461] 195011 212469 272798 236576 188567 118275 181068 229901 275647 222174
#> [471] 172980 120287  93615  95790 132959 115245 103219  96412  86325  77950
#> [481]  98097 131514 118857 111760 100324  91730  81733 102365 127364 113202
#> [491] 108668  96201  84026  73512  87702 105051  93949  88418  76644  63264
#> [501]  52305  60552  76483  69661  64339  52293  44559  37529  41755  54745
#> [511]  52502  51572  44716  33241  34647  37720  45132  39188  36477  30631
#> [521]  27837  25089  31360  44649  44795  46179  43761  41408  38759  49026
#> [531]  68520  70112  73531  70510  67030  62078  75739  99834  93708  91751
#> [541]  82853  75285  67417  81054 109286  99095  94185  82905  72430  61456
#> [551]  68336  92857  80126  73643  54791  47462  37592  42179  53727  49101
#> [561]  44750  38709  32988  27885  30952  37371  33771  31685  26639  21617
#> [571]  19925  20144  25223  25772  21457  18464  15703  13210  14714  17109
#> [581]  14560  13055  11608  10012   8354   9034  11907  13318  11672  10221
#> [591]   8954   7517   9062  10702   9748   8725   7642   6942   6120   7973
#> [601]   9417   8286   7638   6820   5821   5019   5997   7313   6679   6293
#> [611]   5753   5409   4988   5680   7115   7093   7011   6209   6635   7492
#> [621]   9815  11747  11411  11784  11319  10784  10077  12693  15272  15141
#> [631]  14837  14225  13919  13623  16997  19942  20896  20873  19590  18350
#> [641]  16800  20778  25104  25041  24601  24240  23650  21961  27780  33704
#> [651]  31415  29698  26454  24611  21246  25096  29826  26946  23856  21100
#> [661]  18359  15017  17236  18718  16670  16169  14158  12378   9934  11275
#> [671]  13299  11725  10544   9329   8508   7325   8426   9835   9170   8353
#> [681]   7296   6548   5420   6547   8068   7309   6748   6193   5339   4495
#> [691]   5363   6424   5559   5032   4564   4012   3502   4085   5224   4778
#> [701]   4554   3796   3451   3112   3439   4662   5226   4616   4350   3808
#> [711]   3417   4288   5125   4578   4235   3747   3553   3268   4347   5308
#> [721]   5148   4973   4634   4134   3521   4339   6166   7828   7553   6815
#> [731]   6485   5719   7240   8618   8255   8414   7895   7615   7065   9204
#> [741]  11692  10965  10231   9389   8950   7053   8558  10497   9461   8769
#> [751]   8134   7970   6442   7780  10005   8235   7598   6576   6297   4968
#> [761]   5523   6807   5665   5449   4855   4401   3583   4271   5129   4549
#> [771]   4014   3678   3107   2665   3267   4321   3829   3576   3205   2893
#> [781]   2433   2912   3923   3534   3260   3037   2773   2431   2890   3938
#> [791]   3805   3616   3329   2902   2551   3448   4639   4386   3980   3505
#> [801]   3596   2748   3786   5325   5241   5195   4537   4216   3570   4631
#> [811]   6986   7111   6533   6094   6084   4743   5772   9073   8631   7802
#> [821]   6877   5741   4132   3694   4950   6605   8419   7718   6126   4287
#> 
#> 
#> attr(,"class")
#> [1] "EpiStrainDynamics.pathogen_structure"
#> [2] "list"                                

# Using a time series object (time argument is optional)
if (FALSE) { # \dontrun{
sarscov2_xts <- xts::xts(sarscov2[, c("cases", "alpha")], order.by = sarscov2$date)
single(
  data = sarscov2_xts,
  case_timeseries = 'cases'
)
} # }