Subtyped pathogen structure

Subtyped pathogen structure

Usage

subtyped(
  data,
  case_timeseries,
  time,
  influenzaA_unsubtyped_timeseries,
  influenzaA_subtyped_timeseries,
  other_pathogen_timeseries
)

Arguments

data

dataframe containing columns with all relevant data

case_timeseries

name of column with total case data, this name will be used as the name of the pathogen

time

name of column with time data. flexible format - can be date, index, or others, accepted as index identifiers in the tsibble time format.

influenzaA_unsubtyped_timeseries

vector of column names with additional unsubtyped influenzaA case count timeseries

influenzaA_subtyped_timeseries

vector of column names with additional subtyped influenzaA case count timeseries

other_pathogen_timeseries

vector of column names with additional pathogen case count timeseries to model

Value

named list including pathogen_structure, pathogen_names, and data of class EpiStrainDynamics.pathogen_structure

See also

Examples

subtyped(
  data = influenza,
  case_timeseries = 'ili',
  time = 'week',
  influenzaA_unsubtyped_timeseries = 'inf_A',
  influenzaA_subtyped_timeseries = c('inf_H3N2', 'inf_H1N1'),
  other_pathogen_timeseries = c('inf_B', 'other')
)
#> $pathogen_structure
#> [1] "subtyped"
#> 
#> $pathogen_names
#> [1] "inf_H3N2" "inf_H1N1" "inf_B"    "other"   
#> 
#> $data
#> $data$case_timeseries
#>   [1]  29  38  42  58  45  60  30  52  33  26  75  65  71  92  71  75  73  39
#>  [19]  57  94 140 144 218 234 176 265 370 389 469 465 434 496 441 399 359 255
#>  [37] 198 195 156 105  78  48  48  77  69  58  64  80  69  60  50  63  41  54
#>  [55]  64  55  53  64  61  68  81  66  81  89  76 100  86 106  39  57  65  64
#>  [73]  80  83 136 125 140 113 133 171 168 166 204 181 245 216 202 236 220 188
#>  [91] 211 157 105  40  58 128  83 106  70  98 106  32  25  45  48  69  60  79
#> [109]  49  49  81  70  84 107 104  99 110 137  92  98  71 124 114  73  67  98
#> [127]  96 136 148 153 179 173 158 213 306 328 381 410 411 336 341 269 186 166
#> [145] 107 121  47 116 110 110 106  75  59  62  65  34  48  61  61  26  11  11
#> [163]  19  22  28  49  40  39  46  48  36  57  62  66 117  99 118 117 131 141
#> [181] 177 202 220 226 231 246 292 327 384 492 375 340 303 242 206 140  80  88
#> [199]  54  55  55  44  31  40  32  28  15  11  13  21  34  17  13  15  22  30
#> [217]  31  38  31  25  28  44  68  57  59  62  68  55  94  91 123  94 110 109
#> [235] 106 116 131 157 178 229 257 315 247 239 260 241 172 118  95  97  73  52
#> [253]  55  49  50  28  38  37  41  20  19  24  34  28  33  22  20  39  38  55
#> [271]  47  41  51  69  57  49  49  51  57  57  58  57  71  89  71  82 114 160
#> [289] 154 186 243 392 464 540 554 475 508 424 374 308 202 171 132 100  93  83
#> [307]  49  72  46  32  37  49  21  19  36  16  42  20  38  50  44  40  58  33
#> [325]  52  53  43  68  74  54  45  64  85 103 101  83  94 110  79 103 109 123
#> [343] 129 120 129 130 146 152 148 153 134  93  85  98  90  76  54  52  54  41
#> [361]  58  46  51  30  14  32  32  45  24  25  34  30  53  56  50  53  73  88
#> [379] 129 160  95 102 207 209 264 320 330 320 285 283 283 280 319 262 274 246
#> [397] 259 273 242 215 225 150 146 101  89  42  55  37  37  41  37  37  39  35
#> [415]  31  19   7  21  26  19  31  37  45  28  38  33
#> 
#> $data$time
#>   [1] "2012-01-02" "2012-01-09" "2012-01-16" "2012-01-23" "2012-01-30"
#>   [6] "2012-02-06" "2012-02-13" "2012-02-20" "2012-02-27" "2012-03-05"
#>  [11] "2012-03-12" "2012-03-19" "2012-03-26" "2012-04-02" "2012-04-09"
#>  [16] "2012-04-16" "2012-04-23" "2012-04-30" "2012-05-07" "2012-05-14"
#>  [21] "2012-05-21" "2012-05-28" "2012-06-04" "2012-06-11" "2012-06-18"
#>  [26] "2012-06-25" "2012-07-02" "2012-07-09" "2012-07-16" "2012-07-23"
#>  [31] "2012-07-30" "2012-08-06" "2012-08-13" "2012-08-20" "2012-08-27"
#>  [36] "2012-09-03" "2012-09-10" "2012-09-17" "2012-09-24" "2012-10-01"
#>  [41] "2012-10-08" "2012-10-15" "2012-10-22" "2012-10-29" "2012-11-05"
#>  [46] "2012-11-12" "2012-11-19" "2012-11-26" "2012-12-03" "2012-12-10"
#>  [51] "2012-12-17" "2012-12-24" "2012-12-31" "2013-01-07" "2013-01-14"
#>  [56] "2013-01-21" "2013-01-28" "2013-02-04" "2013-02-11" "2013-02-18"
#>  [61] "2013-02-25" "2013-03-04" "2013-03-11" "2013-03-18" "2013-03-25"
#>  [66] "2013-04-01" "2013-04-08" "2013-04-15" "2013-04-22" "2013-04-29"
#>  [71] "2013-05-06" "2013-05-13" "2013-05-20" "2013-05-27" "2013-06-03"
#>  [76] "2013-06-10" "2013-06-17" "2013-06-24" "2013-07-01" "2013-07-08"
#>  [81] "2013-07-15" "2013-07-22" "2013-07-29" "2013-08-05" "2013-08-12"
#>  [86] "2013-08-19" "2013-08-26" "2013-09-02" "2013-09-09" "2013-09-16"
#>  [91] "2013-09-23" "2013-09-30" "2013-10-07" "2013-10-14" "2013-10-21"
#>  [96] "2013-10-28" "2013-11-04" "2013-11-11" "2013-11-18" "2013-11-25"
#> [101] "2013-12-02" "2013-12-09" "2013-12-16" "2013-12-23" "2013-12-30"
#> [106] "2014-01-06" "2014-01-13" "2014-01-20" "2014-01-27" "2014-02-03"
#> [111] "2014-02-10" "2014-02-17" "2014-02-24" "2014-03-03" "2014-03-10"
#> [116] "2014-03-17" "2014-03-24" "2014-03-31" "2014-04-07" "2014-04-14"
#> [121] "2014-04-21" "2014-04-28" "2014-05-05" "2014-05-12" "2014-05-19"
#> [126] "2014-05-26" "2014-06-02" "2014-06-09" "2014-06-16" "2014-06-23"
#> [131] "2014-06-30" "2014-07-07" "2014-07-14" "2014-07-21" "2014-07-28"
#> [136] "2014-08-04" "2014-08-11" "2014-08-18" "2014-08-25" "2014-09-01"
#> [141] "2014-09-08" "2014-09-15" "2014-09-22" "2014-09-29" "2014-10-06"
#> [146] "2014-10-13" "2014-10-20" "2014-10-27" "2014-11-03" "2014-11-10"
#> [151] "2014-11-17" "2014-11-24" "2014-12-01" "2014-12-08" "2014-12-15"
#> [156] "2014-12-22" "2014-12-29" "2015-01-05" "2015-01-12" "2015-01-19"
#> [161] "2015-01-26" "2015-02-02" "2015-02-09" "2015-02-16" "2015-02-23"
#> [166] "2015-03-02" "2015-03-09" "2015-03-16" "2015-03-23" "2015-03-30"
#> [171] "2015-04-06" "2015-04-13" "2015-04-20" "2015-04-27" "2015-05-04"
#> [176] "2015-05-11" "2015-05-18" "2015-05-25" "2015-06-01" "2015-06-08"
#> [181] "2015-06-15" "2015-06-22" "2015-06-29" "2015-07-06" "2015-07-13"
#> [186] "2015-07-20" "2015-07-27" "2015-08-03" "2015-08-10" "2015-08-17"
#> [191] "2015-08-24" "2015-08-31" "2015-09-07" "2015-09-14" "2015-09-21"
#> [196] "2015-09-28" "2015-10-05" "2015-10-12" "2015-10-19" "2015-10-26"
#> [201] "2015-11-02" "2015-11-09" "2015-11-16" "2015-11-23" "2015-11-30"
#> [206] "2015-12-07" "2015-12-14" "2015-12-21" "2015-12-28" "2016-01-04"
#> [211] "2016-01-11" "2016-01-18" "2016-01-25" "2016-02-01" "2016-02-08"
#> [216] "2016-02-15" "2016-02-22" "2016-02-29" "2016-03-07" "2016-03-14"
#> [221] "2016-03-21" "2016-03-28" "2016-04-04" "2016-04-11" "2016-04-18"
#> [226] "2016-04-25" "2016-05-02" "2016-05-09" "2016-05-16" "2016-05-23"
#> [231] "2016-05-30" "2016-06-06" "2016-06-13" "2016-06-20" "2016-06-27"
#> [236] "2016-07-04" "2016-07-11" "2016-07-18" "2016-07-25" "2016-08-01"
#> [241] "2016-08-08" "2016-08-15" "2016-08-22" "2016-08-29" "2016-09-05"
#> [246] "2016-09-12" "2016-09-19" "2016-09-26" "2016-10-03" "2016-10-10"
#> [251] "2016-10-17" "2016-10-24" "2016-10-31" "2016-11-07" "2016-11-14"
#> [256] "2016-11-21" "2016-11-28" "2016-12-05" "2016-12-12" "2016-12-19"
#> [261] "2016-12-26" "2017-01-02" "2017-01-09" "2017-01-16" "2017-01-23"
#> [266] "2017-01-30" "2017-02-06" "2017-02-13" "2017-02-20" "2017-02-27"
#> [271] "2017-03-06" "2017-03-13" "2017-03-20" "2017-03-27" "2017-04-03"
#> [276] "2017-04-10" "2017-04-17" "2017-04-24" "2017-05-01" "2017-05-08"
#> [281] "2017-05-15" "2017-05-22" "2017-05-29" "2017-06-05" "2017-06-12"
#> [286] "2017-06-19" "2017-06-26" "2017-07-03" "2017-07-10" "2017-07-17"
#> [291] "2017-07-24" "2017-07-31" "2017-08-07" "2017-08-14" "2017-08-21"
#> [296] "2017-08-28" "2017-09-04" "2017-09-11" "2017-09-18" "2017-09-25"
#> [301] "2017-10-02" "2017-10-09" "2017-10-16" "2017-10-23" "2017-10-30"
#> [306] "2017-11-06" "2017-11-13" "2017-11-20" "2017-11-27" "2017-12-04"
#> [311] "2017-12-11" "2017-12-18" "2017-12-25" "2018-01-01" "2018-01-08"
#> [316] "2018-01-15" "2018-01-22" "2018-01-29" "2018-02-05" "2018-02-12"
#> [321] "2018-02-19" "2018-02-26" "2018-03-05" "2018-03-12" "2018-03-19"
#> [326] "2018-03-26" "2018-04-02" "2018-04-09" "2018-04-16" "2018-04-23"
#> [331] "2018-04-30" "2018-05-07" "2018-05-14" "2018-05-21" "2018-05-28"
#> [336] "2018-06-04" "2018-06-11" "2018-06-18" "2018-06-25" "2018-07-02"
#> [341] "2018-07-09" "2018-07-16" "2018-07-23" "2018-07-30" "2018-08-06"
#> [346] "2018-08-13" "2018-08-20" "2018-08-27" "2018-09-03" "2018-09-10"
#> [351] "2018-09-17" "2018-09-24" "2018-10-01" "2018-10-08" "2018-10-15"
#> [356] "2018-10-22" "2018-10-29" "2018-11-05" "2018-11-12" "2018-11-19"
#> [361] "2018-11-26" "2018-12-03" "2018-12-10" "2018-12-17" "2018-12-24"
#> [366] "2018-12-31" "2019-01-07" "2019-01-14" "2019-01-21" "2019-01-28"
#> [371] "2019-02-04" "2019-02-11" "2019-02-18" "2019-02-25" "2019-03-04"
#> [376] "2019-03-11" "2019-03-18" "2019-03-25" "2019-04-01" "2019-04-08"
#> [381] "2019-04-15" "2019-04-22" "2019-04-29" "2019-05-06" "2019-05-13"
#> [386] "2019-05-20" "2019-05-27" "2019-06-03" "2019-06-10" "2019-06-17"
#> [391] "2019-06-24" "2019-07-01" "2019-07-08" "2019-07-15" "2019-07-22"
#> [396] "2019-07-29" "2019-08-05" "2019-08-12" "2019-08-19" "2019-08-26"
#> [401] "2019-09-02" "2019-09-09" "2019-09-16" "2019-09-23" "2019-09-30"
#> [406] "2019-10-07" "2019-10-14" "2019-10-21" "2019-10-28" "2019-11-04"
#> [411] "2019-11-11" "2019-11-18" "2019-11-25" "2019-12-02" "2019-12-09"
#> [416] "2019-12-16" "2019-12-23" "2019-12-30" "2020-01-06" "2020-01-13"
#> [421] "2020-01-20" "2020-01-27" "2020-02-03" "2020-02-10" "2020-02-17"
#> [426] "2020-02-24"
#> 
#> $data$component_pathogens
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#> inf_A    4    4   13    6    3    1    2    4    6     4    12     9    14
#> inf_B    3    1    8    5    2    7   10    9   13    10    19    12     8
#> other  278  219  238  187  228  274  278  266  327   324   346   333   322
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#> inf_A     8     9    16    13    14    12    15    27    33    56    95    95
#> inf_B     5     6     4     4    10     7    12    11     8    18    23    30
#> other   317   326   378   328   379   403   481   524   637   509   512   701
#>       [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37]
#> inf_A   182   264   310   280   311   318   274   295   187   153    89    63
#> inf_B    72    90   129   111   122   117   114    90   100    97    50    46
#> other   644   675   949   903   775   874   844   783   897   686   813   707
#>       [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49]
#> inf_A    44    17    12    10     9     4     0    10    10     0     3     7
#> inf_B    28    15    12     7     6     3     2     3     1     2     1     0
#> other   508   524   450   433   433   367   324   352   374   359   319   357
#>       [,50] [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61]
#> inf_A     2     3     1     5     8    11    11    10    15    15     8    15
#> inf_B     0     2     0     3     5     6     4     3     3     2     2     1
#> other   292   258   205   153   177   171   171   185   198   198   240   246
#>       [,62] [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73]
#> inf_A     8    18    25    10     6     4     4     3     1     2     4     9
#> inf_B     7     3     2     0     2     1     0     0     0     2     3     1
#> other   260   285   305   250   248   289   272   303   296   284   282   378
#>       [,74] [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85]
#> inf_A     5     8     4     4     7    23    26    48    37    71    87    80
#> inf_B     0     1     3     1     3     6    10    20     8    10    17    19
#> other   339   317   393   330   370   425   423   441   471   459   469   507
#>       [,86] [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97]
#> inf_A   105   127   102   116   113   109    59    64    46    38    39    26
#> inf_B    23    32    14    24    12    16    14    18     7     4     8     6
#> other   566   561   611   509   586   549   461   430   467   321   322   277
#>       [,98] [,99] [,100] [,101] [,102] [,103] [,104] [,105] [,106] [,107]
#> inf_A    17    10     19     12     22      7     11     20     22     23
#> inf_B     8     4     13      5      9      1      2      1      1      2
#> other   269   244    254    217    229    239    119    142    170    173
#>       [,108] [,109] [,110] [,111] [,112] [,113] [,114] [,115] [,116] [,117]
#> inf_A     17     18     29     25     13     20     18     18     10     14
#> inf_B      9      1      2      1      3     13      3      3      7      2
#> other    201    160    179    122    197    237    152    254    211    211
#>       [,118] [,119] [,120] [,121] [,122] [,123] [,124] [,125] [,126] [,127]
#> inf_A     15     12     16      6      5     12     12     15     13     12
#> inf_B      6      7      1      2      4      4      7      4      1      2
#> other    280    229    232    238    231    294    269    310    309    280
#>       [,128] [,129] [,130] [,131] [,132] [,133] [,134] [,135] [,136] [,137]
#> inf_A     16     19     44     83    118    172    176    204    268    237
#> inf_B      3      4      7      8      7     12     12     13      9     31
#> other    329    316    339    374    435    481    451    508    527    631
#>       [,138] [,139] [,140] [,141] [,142] [,143] [,144] [,145] [,146] [,147]
#> inf_A    293    246    187    162    128     92     51     41     17     17
#> inf_B     27     25     30     25     22     18     21     17     12     11
#> other    640    548    532    541    532    524    398    424    357    374
#>       [,148] [,149] [,150] [,151] [,152] [,153] [,154] [,155] [,156] [,157]
#> inf_A     14      7     21      1      2      6      9      5     10      2
#> inf_B     10     18      5      7      4      6      6      2      4      1
#> other    354    312    294    312    247    216    232    223    146     68
#>       [,158] [,159] [,160] [,161] [,162] [,163] [,164] [,165] [,166] [,167]
#> inf_A     19     19     15     11     22     17     25     17     17     17
#> inf_B      1      4      3      5      2      5      5      3      5      3
#> other    197    169    177    175    172    175    216    213    231    214
#>       [,168] [,169] [,170] [,171] [,172] [,173] [,174] [,175] [,176] [,177]
#> inf_A     14     19     32     13     15     28     23     10     17     21
#> inf_B      5      9      5      9     10     11     18     15     25     26
#> other    248    236    267    226    274    291    256    301    295    330
#>       [,178] [,179] [,180] [,181] [,182] [,183] [,184] [,185] [,186] [,187]
#> inf_A     12     22     25     23     23     35     42     40     49     61
#> inf_B     29     41     46     49     59    111     88     92    106     93
#> other    349    349    373    421    403    456    485    502    542    592
#>       [,188] [,189] [,190] [,191] [,192] [,193] [,194] [,195] [,196] [,197]
#> inf_A     78     72    114     74     91     86    102    117     68     58
#> inf_B    125    103    137    111     79     73     79     72     45     27
#> other    549    619    624    622    563    560    549    491    385    387
#>       [,198] [,199] [,200] [,201] [,202] [,203] [,204] [,205] [,206] [,207]
#> inf_A     49     27     19     24     31      9      9     22     29     15
#> inf_B     19     14      8      6      4      3      2      2      2      2
#> other    338    321    323    314    277    286    275    295    265    254
#>       [,208] [,209] [,210] [,211] [,212] [,213] [,214] [,215] [,216] [,217]
#> inf_A     20      6     27     17     11     16     21     22     16     24
#> inf_B      3      0      0      2      1      1      2      8      3      3
#> other    183    108    253    223    207    197    212    222    221    250
#>       [,218] [,219] [,220] [,221] [,222] [,223] [,224] [,225] [,226] [,227]
#> inf_A     21     22     16     21      7     14     15      8      6     14
#> inf_B      8      4      6      7      5      6      8      4      5      5
#> other    273    258    267    241    269    284    316    319    306    367
#>       [,228] [,229] [,230] [,231] [,232] [,233] [,234] [,235] [,236] [,237]
#> inf_A     19     30     26     35     39     52     63     92     99    121
#> inf_B     12     14      7     12      9     12     14     24     33     55
#> other    389    405    860    805    746    788    897    950   1052   1075
#>       [,238] [,239] [,240] [,241] [,242] [,243] [,244] [,245] [,246] [,247]
#> inf_A    239    243    284    342    488    479    541    539    372    317
#> inf_B     41     54     67     91     81     75     70     63     61     62
#> other   1075   1078   1235   1227   1456   1367   1572   1587   1404   1409
#>       [,248] [,249] [,250] [,251] [,252] [,253] [,254] [,255] [,256] [,257]
#> inf_A    236    185    128     65     46     48     20     22     27     12
#> inf_B     37     41     27     17     15     12     14      4      5      8
#> other   1255   1250   1163   1017    954    877    848    722    634    607
#>       [,258] [,259] [,260] [,261] [,262] [,263] [,264] [,265] [,266] [,267]
#> inf_A     13     17     20      9     30     30     25     37     49     40
#> inf_B      6      4     10      4      6      2      3      2      2      1
#> other    582    557    522    429    326    352    309    275    263    231
#>       [,268] [,269] [,270] [,271] [,272] [,273] [,274] [,275] [,276] [,277]
#> inf_A     30     42     25     18     30     34     32     15     17      7
#> inf_B      3      8      2      2      2      2      8      6      1      4
#> other    288    316    381    344    310    509    652    596    593    734
#>       [,278] [,279] [,280] [,281] [,282] [,283] [,284] [,285] [,286] [,287]
#> inf_A     10     21     16     28     26     20     24     32     61     59
#> inf_B      4      6      3      4     15      7      7     19     27     23
#> other    675    717    719    725    789    905    781    842    989    947
#>       [,288] [,289] [,290] [,291] [,292] [,293] [,294] [,295] [,296] [,297]
#> inf_A    125    225    363    359    471    610    652    654    584    709
#> inf_B     45     59     74     69    148    207    249    222    220    251
#> other   1114   1238   1150   1283   1137   1088   1653   1765   1796   1832
#>       [,298] [,299] [,300] [,301] [,302] [,303] [,304] [,305] [,306] [,307]
#> inf_A    663    502    329    221    132     76     48     40     29     29
#> inf_B    206    219    200    126     96     52     30     36     33     20
#> other   1927   1817   1691   1488   1377   1087   1058   1028    987    929
#>       [,308] [,309] [,310] [,311] [,312] [,313] [,314] [,315] [,316] [,317]
#> inf_A     13     21     19     17     17     18     38     19     24     34
#> inf_B     22     22     11     15     13     11      9     11     12     22
#> other    746    752    671    641    606    525    428    429    364    337
#>       [,318] [,319] [,320] [,321] [,322] [,323] [,324] [,325] [,326] [,327]
#> inf_A     36     26     55     37     23     27     16     18     22     23
#> inf_B     17     10     18     16     14      5      6     10      3      7
#> other    384    326    336    331    368    373    345    544    504    621
#>       [,328] [,329] [,330] [,331] [,332] [,333] [,334] [,335] [,336] [,337]
#> inf_A     14     20     16     26     28     29     18     20     25     37
#> inf_B      0      3      2      5      4      4      6      4      0      7
#> other    678    702    669    749    815    933    980    945   1094   1085
#>       [,338] [,339] [,340] [,341] [,342] [,343] [,344] [,345] [,346] [,347]
#> inf_A     39     70     64     89     70     80     71    114    145    186
#> inf_B      2      2      3      7      3      8      0      5      9     14
#> other   1188   1163   1244   1427   1422   1408   1399   1341   1476   1457
#>       [,348] [,349] [,350] [,351] [,352] [,353] [,354] [,355] [,356] [,357]
#> inf_A    201    155    238    232    158    176    105     74     56     39
#> inf_B      9     19     14      8     12      7     13      9      3      4
#> other   1440   1719   1412   1469   1332   1360   1234   1076    941    936
#>       [,358] [,359] [,360] [,361] [,362] [,363] [,364] [,365] [,366] [,367]
#> inf_A     47     63    105    110    186    124    138    103     84     99
#> inf_B      6      7      8      7      3      5      7      6      3      3
#> other    960    943    933    984    843    857    828    664    646    790
#>       [,368] [,369] [,370] [,371] [,372] [,373] [,374] [,375] [,376] [,377]
#> inf_A    123     95    107    104    103    124    132     95    105     67
#> inf_B      2      2      3      2      1      1      2      2      6      7
#> other    717    688    778    669    784    876    887    868    884    983
#>       [,378] [,379] [,380] [,381] [,382] [,383] [,384] [,385] [,386] [,387]
#> inf_A     76     94    142    174    196    213    245    363    640    655
#> inf_B      5      4      2      4     14      9     12     22     46     52
#> other    892   1095   1176   1129   1271   1267   1327   1556   2041   2178
#>       [,388] [,389] [,390] [,391] [,392] [,393] [,394] [,395] [,396] [,397]
#> inf_A    844    746    782    862    749    567    511    364    294    303
#> inf_B     68     96    109    140    118     95     96     97     87    118
#> other   2207   2478   2433   2563   2671   2712   2573   2297   2245   2294
#>       [,398] [,399] [,400] [,401] [,402] [,403] [,404] [,405] [,406] [,407]
#> inf_A    274    338    256    182    159    130    127     79     78     52
#> inf_B    154    138    103    112     65     51     32     28     22      4
#> other   2350   2239   2247   2202   1882   2011   1923   1767   1725   1518
#>       [,408] [,409] [,410] [,411] [,412] [,413] [,414] [,415] [,416] [,417]
#> inf_A     38     29     20     25     23     36     51     37     18     25
#> inf_B      7      5      2      0      3      4      3      5      0      1
#> other   1563   1367   1226   1219   1202   1107   1160   1142    805    852
#>       [,418] [,419] [,420] [,421] [,422] [,423] [,424] [,425] [,426]
#> inf_A     41     58     54    103     72     74     62     55     69
#> inf_B      1      2      2     10     12      6      5      7      3
#> other    979   1059    891   1057   1290   1290   1193   1293   1513
#> 
#> $data$influenzaA_subtyped
#>          [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#> inf_H3N2    4    4   13    3    3    1    0    3    4     4    10     9    14
#> inf_H1N1    0    0    0    2    0    0    2    1    2     0     2     0     0
#>          [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24]
#> inf_H3N2     7     6    13    12    13    11    13    26    28    39    63
#> inf_H1N1     0     3     3     1     1     1     2     1     2     1     1
#>          [,25] [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35]
#> inf_H3N2    79   175   250   306   272   304   308   268   291   180   148
#> inf_H1N1     2     5    11     4     7     5     9     6     3     6     4
#>          [,36] [,37] [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46]
#> inf_H3N2    88    61    43    15    12     8     9     4     0     8     8
#> inf_H1N1     1     2     0     0     0     2     0     0     0     0     1
#>          [,47] [,48] [,49] [,50] [,51] [,52] [,53] [,54] [,55] [,56] [,57]
#> inf_H3N2     0     1     4     1     1     0     3     2     7     8     6
#> inf_H1N1     0     2     3     0     2     1     2     5     3     3     4
#>          [,58] [,59] [,60] [,61] [,62] [,63] [,64] [,65] [,66] [,67] [,68]
#> inf_H3N2    11     9     7    13     6     7     7     7     4     2     2
#> inf_H1N1     3     6     1     1     2    11    18     3     1     2     2
#>          [,69] [,70] [,71] [,72] [,73] [,74] [,75] [,76] [,77] [,78] [,79]
#> inf_H3N2     1     1     1     4     7     3     4     3     4     5    16
#> inf_H1N1     2     0     1     0     1     1     4     1     0     2     5
#>          [,80] [,81] [,82] [,83] [,84] [,85] [,86] [,87] [,88] [,89] [,90]
#> inf_H3N2    19    40    28    58    65    61    74    70    49    75    70
#> inf_H1N1     7     7     8    12    21    18    30    55    52    40    42
#>          [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98] [,99] [,100] [,101]
#> inf_H3N2    65    29    32    30    21    22    14    10     6     12      8
#> inf_H1N1    42    30    30    15    17    14    11     7     3      5      4
#>          [,102] [,103] [,104] [,105] [,106] [,107] [,108] [,109] [,110] [,111]
#> inf_H3N2     21      6     10     14     15     12     10     14     24     18
#> inf_H1N1      1      1      1      5      7     11      7      4      5      7
#>          [,112] [,113] [,114] [,115] [,116] [,117] [,118] [,119] [,120] [,121]
#> inf_H3N2      9     11     12      7      2      8      9      7     15      5
#> inf_H1N1      3      9      6     11      8      6      6      4      1      1
#>          [,122] [,123] [,124] [,125] [,126] [,127] [,128] [,129] [,130] [,131]
#> inf_H3N2      4      6      3      4      6      2      1      3     12     25
#> inf_H1N1      1      5      9      9      7     10     13     14     30     54
#>          [,132] [,133] [,134] [,135] [,136] [,137] [,138] [,139] [,140] [,141]
#> inf_H3N2     14     33     48     82    110     81     91     73     65     52
#> inf_H1N1    104    131    125    110    148    136    168    152    104     98
#>          [,142] [,143] [,144] [,145] [,146] [,147] [,148] [,149] [,150] [,151]
#> inf_H3N2     57     36     27     21     16     11      7      3     12      1
#> inf_H1N1     60     46     18     17      1      6      7      4      8      0
#>          [,152] [,153] [,154] [,155] [,156] [,157] [,158] [,159] [,160] [,161]
#> inf_H3N2      2      3      7      4     10      2     17     16     14      8
#> inf_H1N1      0      2      2      1      0      0      2      3      1      3
#>          [,162] [,163] [,164] [,165] [,166] [,167] [,168] [,169] [,170] [,171]
#> inf_H3N2     18     14     20     11     14     17     11     14     29     12
#> inf_H1N1      3      3      5      6      3      0      2      5      2      1
#>          [,172] [,173] [,174] [,175] [,176] [,177] [,178] [,179] [,180] [,181]
#> inf_H3N2     13     22     19     10     14     18     10     21     17     23
#> inf_H1N1      2      6      4      0      3      2      2      1      7      0
#>          [,182] [,183] [,184] [,185] [,186] [,187] [,188] [,189] [,190] [,191]
#> inf_H3N2     19     32     36     36     38     52     63     63     98     64
#> inf_H1N1      3      3      6      4      7      8     11      6     12      7
#>          [,192] [,193] [,194] [,195] [,196] [,197] [,198] [,199] [,200] [,201]
#> inf_H3N2     80     73     92    103     54     51     44     20     15     23
#> inf_H1N1      8      7      6     12     10      5      5      5      4      1
#>          [,202] [,203] [,204] [,205] [,206] [,207] [,208] [,209] [,210] [,211]
#> inf_H3N2     30      9      8     17     13      4      8      4     16      3
#> inf_H1N1      1      0      1      5     16     11     12      2     11     14
#>          [,212] [,213] [,214] [,215] [,216] [,217] [,218] [,219] [,220] [,221]
#> inf_H3N2      5      2      4      2      3      6      0      4      4     11
#> inf_H1N1      6     14     17     20     13     18     21     18     12     10
#>          [,222] [,223] [,224] [,225] [,226] [,227] [,228] [,229] [,230] [,231]
#> inf_H3N2      3      2      4      4      3      9      8     12      7      6
#> inf_H1N1      4     12     11      4      3      5     11     18     14     18
#>          [,232] [,233] [,234] [,235] [,236] [,237] [,238] [,239] [,240] [,241]
#> inf_H3N2      9     10     19     22     22     41     94    131    188    221
#> inf_H1N1     13     26     17     26     25     21     37     17     24     17
#>          [,242] [,243] [,244] [,245] [,246] [,247] [,248] [,249] [,250] [,251]
#> inf_H3N2    242    236    293    291    216    196    146    118     94     43
#> inf_H1N1     17     14     11     10     11      7      7      6      5      0
#>          [,252] [,253] [,254] [,255] [,256] [,257] [,258] [,259] [,260] [,261]
#> inf_H3N2     26     32     12     13     15      8      6     10     18      3
#> inf_H1N1      1      0      0      0      0      1      0      1      0      0
#>          [,262] [,263] [,264] [,265] [,266] [,267] [,268] [,269] [,270] [,271]
#> inf_H3N2     23     19     18     22     37     32     21     34     21     14
#> inf_H1N1      2      2      1      3      2      0      0      1      0      2
#>          [,272] [,273] [,274] [,275] [,276] [,277] [,278] [,279] [,280] [,281]
#> inf_H3N2     19     18     15      9      6      2      4      9      6      8
#> inf_H1N1      2      5      3      1      2      4      0      8      7     11
#>          [,282] [,283] [,284] [,285] [,286] [,287] [,288] [,289] [,290] [,291]
#> inf_H3N2      7      8      9     12     13     11     29     54     59     57
#> inf_H1N1      8      5      4      8     15     17     15     18     18     13
#>          [,292] [,293] [,294] [,295] [,296] [,297] [,298] [,299] [,300] [,301]
#> inf_H3N2     65     95    115    142    135    238    217    167    109     93
#> inf_H1N1     16     14     30     18     26     34     37     34     33     28
#>          [,302] [,303] [,304] [,305] [,306] [,307] [,308] [,309] [,310] [,311]
#> inf_H3N2     69     52     36     31     19     16      6      9     12      6
#> inf_H1N1     17      4      4      6      2      3      5      6      3      4
#>          [,312] [,313] [,314] [,315] [,316] [,317] [,318] [,319] [,320] [,321]
#> inf_H3N2      7     13     13      6     12     10     20     12     26     18
#> inf_H1N1      5      1     14      6      5      8      4      4     13      4
#>          [,322] [,323] [,324] [,325] [,326] [,327] [,328] [,329] [,330] [,331]
#> inf_H3N2     13     14      8      8      8      6      6      3      5      5
#> inf_H1N1      5      4      7      8     11      8      5      7      8     10
#>          [,332] [,333] [,334] [,335] [,336] [,337] [,338] [,339] [,340] [,341]
#> inf_H3N2     12      6      1      2      8      3      2      6      9      7
#> inf_H1N1      6      6      5     11     12     27     24     41     48     66
#>          [,342] [,343] [,344] [,345] [,346] [,347] [,348] [,349] [,350] [,351]
#> inf_H3N2      5      5      4      8      7     15     14     23     26     25
#> inf_H1N1     44     44     44     68     92     93    112    100    107    144
#>          [,352] [,353] [,354] [,355] [,356] [,357] [,358] [,359] [,360] [,361]
#> inf_H3N2     29     55     29     18     24      7      8     13      5     14
#> inf_H1N1     81     61     46     26     21     21     25     39     78     73
#>          [,362] [,363] [,364] [,365] [,366] [,367] [,368] [,369] [,370] [,371]
#> inf_H3N2     28     23     28     33     29     39     60     47     23     47
#> inf_H1N1    136     91     92     43     30     51     53     26     29     28
#>          [,372] [,373] [,374] [,375] [,376] [,377] [,378] [,379] [,380] [,381]
#> inf_H3N2     33     45     48     23     29     13     17     24     73     84
#> inf_H1N1     28     29     26     25     21     15     13     21     13     24
#>          [,382] [,383] [,384] [,385] [,386] [,387] [,388] [,389] [,390] [,391]
#> inf_H3N2    108    104    162    204    390    394    471    365    345    355
#> inf_H1N1     17     21     12     19     15      6      9      5     10      3
#>          [,392] [,393] [,394] [,395] [,396] [,397] [,398] [,399] [,400] [,401]
#> inf_H3N2    314    178    150    110     60     40     33     52     18     27
#> inf_H1N1      4      8      6      6      3      3      4      6      9      5
#>          [,402] [,403] [,404] [,405] [,406] [,407] [,408] [,409] [,410] [,411]
#> inf_H3N2     23      7      9      5      6      9      7      1      0      2
#> inf_H1N1      1      1      1      3      2      2      4      4      5      6
#>          [,412] [,413] [,414] [,415] [,416] [,417] [,418] [,419] [,420] [,421]
#> inf_H3N2      1      0      1      1      0      0      3      3      2      2
#> inf_H1N1      6      9     13     14      0      0     11     11     21     37
#>          [,422] [,423] [,424] [,425] [,426]
#> inf_H3N2      5      3      2      7     11
#> inf_H1N1     28     22     16     17     22
#> 
#> 
#> attr(,"class")
#> [1] "EpiStrainDynamics.pathogen_structure"
#> [2] "list"