Generic Method for fitting model

S3 generic for fitted models from constructed model object describe how to set up inits.

Usage

fit_model(
  constructed_model,
  n_chain = 4,
  n_iter = 2000,
  n_warmup = floor(n_iter/2),
  thin = 1,
  adapt_delta = 0.9,
  multi_cores = TRUE,
  verbose = TRUE,
  seed = NULL,
  ...
)

# S3 method for class 'rw_subtyped'
fit_model(
  constructed_model,
  n_chain = 4,
  n_iter = 2000,
  n_warmup = floor(n_iter/2),
  thin = 1,
  adapt_delta = 0.9,
  multi_cores = TRUE,
  verbose = TRUE,
  seed = NULL,
  ...
)

# S3 method for class 'ps_subtyped'
fit_model(
  constructed_model,
  n_chain = 4,
  n_iter = 2000,
  n_warmup = floor(n_iter/2),
  thin = 1,
  adapt_delta = 0.9,
  multi_cores = TRUE,
  verbose = TRUE,
  seed = NULL,
  ...
)

# S3 method for class 'rw_multiple'
fit_model(
  constructed_model,
  n_chain = 4,
  n_iter = 2000,
  n_warmup = floor(n_iter/2),
  thin = 1,
  adapt_delta = 0.9,
  multi_cores = TRUE,
  verbose = TRUE,
  seed = NULL,
  ...
)

# S3 method for class 'ps_multiple'
fit_model(
  constructed_model,
  n_chain = 4,
  n_iter = 2000,
  n_warmup = floor(n_iter/2),
  thin = 1,
  adapt_delta = 0.9,
  multi_cores = TRUE,
  verbose = TRUE,
  seed = NULL,
  ...
)

# S3 method for class 'rw_single'
fit_model(
  constructed_model,
  n_chain = 4,
  n_iter = 2000,
  n_warmup = floor(n_iter/2),
  thin = 1,
  adapt_delta = 0.9,
  multi_cores = TRUE,
  verbose = TRUE,
  seed = NULL,
  ...
)

# S3 method for class 'ps_single'
fit_model(
  constructed_model,
  n_chain = 4,
  n_iter = 2000,
  n_warmup = floor(n_iter/2),
  thin = 1,
  adapt_delta = 0.9,
  multi_cores = TRUE,
  verbose = TRUE,
  seed = NULL,
  ...
)

Arguments

constructed_model

prepared model object of class EpiStrainDynamics.model

n_chain

number of MCMC chains, defaults to 4

n_iter

A positive integer specifying the number of iterations for each chain, default value is 2000

n_warmup

A positive integer specifying the number of warmup iterations, default value is half the number of iterations

thin

A positive integer specifying the period for saving samples, default value is 1.

adapt_delta

Numeric value between 0 and 1 indicating target average acceptance probability used in rstan::sampling. Default value is 0.8.

multi_cores

A logical value indicating whether to parallelize chains with multiple cores, default is TRUE and uses all available cores - 1.

verbose

Logical value controlling the verbosity of output (i.e., warnings, messages, progress bar), default is TRUE.

seed

A positive integer seed used for random number generation in MCMC. Default is NULL, which means the seed is generated from 1 to the maximum integer supported by R.

...

additional arguments to rstan::sampling()

Value

fit model of class EpiStrainDynamics.fit

Examples

if (FALSE) { # \dontrun{
  mod <- construct_model(
    method = random_walk(),
    pathogen_structure = single(
      case_timeseries = sarscov2$cases,
      time = sarscov2$date))

  fit <- fit_model(mod)

  # or specify additional mcmc parameters
  fit <- fit_model(
    mod, iter = 3000, warmup = 2000, chains = 4
  )
} # }